Fix clippy findings across the workspace, so a clippy gate can be strict
CI is about to enforce clippy with -D warnings. Clear the findings first:

- `x % n == 0` alignment checks become `x.is_multiple_of(n)` (frames, heap,
  slab, paging), which reads as the intent ("is this frame-aligned").
- The virtio shadow-ring copy had a redundant `as u64` cast on a value
  already u64.
- GIC_REGIONS gets a `GicRegions` type alias (clippy::type_complexity).
- Two nested `if let` blocks collapse into let-chains (sched::wake,
  main's milestone-11 demo), which edition 2024 makes legal.
- The user binary's diverging tails spin with `core::hint::spin_loop()`
  instead of a bare `loop {}`, matching the idiom already used elsewhere.
- Two deliberate spots keep their form under a scoped #[allow] with a
  reason: MAIR's `| 0x00` spells out slot 0, and a test name shouts
  "TABLE" on purpose.
- Dropped a dead `SPAWN_QUOTA` static and an unused `AtomicU32` import
  from two kernel tests.

Behaviour is unchanged; 96 kernel tests still pass. Both clippy command
groups (host crates + xtask, and kernel + user on the aarch64 target) are
clean under -D warnings.

+++ kernel/src/virtio.rs
@@ -272,7 +272,7 @@ fn validate_and_shadow(
             let src = driver_desc + d as u64 * 16;
             let addr = read64(src);
             let word = read64(src + 8);
-            let len = (word & 0xffff_ffff) as u64;
+            let len = word & 0xffff_ffff;
             let flags = ((word >> 32) & 0xffff) as u16;
             let next = ((word >> 48) & 0xffff) as u16;
